Minor change to allow to work on more versions of Linux

+++ bin/copy.sh
@@ -10,23 +10,31 @@
 # granularity.
 
 # Various settings
-volume=/volume1
+#volume=/volume1
+volume=/home/docker/Repositories/backup
 backupDir=$volume/Backups
 archiveDir=${backupDir}/Archives
 tempArchive=${backupDir}/temp-archive.tar.gz
 dataDir=${backupDir}/data
 logDir=${backupDir}/logs
 logfile=$logDir/$(date -I).log
+gnuDir=${backupDir}/gnupg
+
+# Check that our volume exists
+if [[ ! -e $volume ]] ; then
+    echo "FATAL: Volume not found: $volume"
+    exit 1
+fi
 
 # We need a log directory
 if [[ ! -e $logDir ]] ; then
     mkdir -p $logDir
 fi
-echo "[$(date -Ins)] INFO Started backup run" >> $logfile
+echo "[$(date -Is)] INFO Started backup run" >> $logfile
 
 # If we don't have a data directory, create one
 if [[ ! -e $dataDir ]] ; then
-	echo "[$(date -Ins)] INFO No data directory, so created it" >> $logfile
+	echo "[$(date -Is)] INFO No data directory, so created it" >> $logfile
     mkdir -p $dataDir
     echo "#!/bin/sh" > $dataDir/folders.sh.template
     echo "folders=(" >> $dataDir/folders.sh.template
@@ -35,6 +43,13 @@ if [[ ! -e $dataDir ]] ; then
     echo "    homes/example/Photos:0" >> $dataDir/folders.sh.template
 	echo ")" >> $dataDir/folders.sh.template
 fi
+
+# We want a temporary directory for our gpg internals.  If there is already
+# one, delete it, then create a new one
+if [[ -e $gnuDir ]] ; then
+    rm -rf $gnuDir
+fi
+mkdir $gnuDir
 
 # The folders we want to backup.  These are in the format of a directory path
 # and a number.  The directory path is the folder to be backed up, and the
@@ -54,6 +69,11 @@ fi
 #     )
 #
 # For separation, we keep this in another file
+if [[ ! -e $dataDir/folders.sh ]] ; then
+    echo "FATAL: List of folders to backup does not exist: $dataDir/folders.sh"
+    echo "[$(date -Is)] FATAL: List of folders to backup does not exist: $dataDir/folders.sh" >> $logfile
+    exit 1
+fi
 source $dataDir/folders.sh
 
 # Go through the folders, identifying each of the folders which need to be
@@ -62,7 +82,7 @@ for folder in ${folders[*]} ; do
 
     path=$(echo $folder | cut -d':' -f1)
     depth=$(echo $folder | cut -d':' -f2)
-    echo "[$(date -Ins)] INFO Analysing path ${path}, depth ${depth}" >> $logfile
+    echo "[$(date -Is)] INFO Analysing path ${path}, depth ${depth}" >> $logfile
 
     # Set both max and min depth to be our desired depth, so we only find
     # directories at the desired depth
@@ -72,7 +92,7 @@ for folder in ${folders[*]} ; do
 
         # Ignore any NAS-specific directories
         if [[ "$sourceFileset" != "@eaDir" ]] ; then
-            echo "[$(date -Ins)] INFO Identified directory for backup: ${line}" >> $logfile
+            echo "[$(date -Is)] INFO Identified directory for backup: ${line}" >> $logfile
 
             # Derive a more unix-y name for the archive file, from the directory name
             archive=$(echo "$sourceFileset" | \
@@ -82,7 +102,7 @@ for folder in ${folders[*]} ; do
             # If the backup directory doesn't exist, create it
             targetDir=$archiveDir$(echo "$sourceDirectory" | sed -e "s|^$volume||")
             if [[ ! -e $targetDir ]] ; then
-                echo "[$(date -Ins)] INFO Creating directory: ${targetDir}" >> $logfile
+                echo "[$(date -Is)] INFO Creating directory: ${targetDir}" >> $logfile
                 mkdir -p ${targetDir} >> $logfile
             fi
 
@@ -90,29 +110,29 @@ for folder in ${folders[*]} ; do
             performBackup="no"
             if [[ ! -e $targetDir/$archive ]] ; then
                 performBackup="yes"
-                echo "[$(date -Ins)] INFO Archive file does not exist: ${targetDir}/${archive}" >> $logfile
+                echo "[$(date -Is)] INFO Archive file does not exist: ${targetDir}/${archive}" >> $logfile
             fi
 
             # Check for changes made since the last backup
-            echo "[$(date -Ins)] WARNING Missing code here!" >> $logfile
+            echo "[$(date -Is)] WARNING Missing code here!" >> $logfile
 
             # Actually perform the backup (if needed).  To avoid getting
             # partial archives (if the script crashes for some reason, or gets
             # interrupted), we create the archive as a temporary file, then
             # rename it to the proper location when (if) it completes successfully
             if [[ "$performBackup" == "yes" ]] ; then
-                echo "[$(date -Ins)] INFO Creating archive: ${targetDir}/${archive}" >> $logfile
+                echo "[$(date -Is)] INFO Creating archive: ${targetDir}/${archive}" >> $logfile
                 tar zcvf $tempArchive -C "$sourceDirectory" --exclude='@eaDir' "$sourceFileset" >> $logfile
                 tarStatus=$?
                 if [[ $tarStatus -eq 0 ]] ; then
                     cp $tempArchive "${targetDir}/${archive}"
                     rm $tempArchive
                 else
-                    echo "[$(date -Ins)] ERROR Could not create archive: ${targetDir}/${archive}, error: $tarStatus" >> $logfile
+                    echo "[$(date -Is)] ERROR Could not create archive: ${targetDir}/${archive}, error: $tarStatus" >> $logfile
                 fi
             fi
         fi
     done
 done
 
-echo "[$(date -Ins)] INFO Finished backup run" >> $logfile
+echo "[$(date -Is)] INFO Finished backup run" >> $logfile
